fix: move grapqh queries mutations in separate folder
